Made type of MaxCanBoards consistent between boards

+++ src/Config/Pins_Duet3Mini.h
@@ -82,7 +82,7 @@ constexpr size_t MaxSmartDrivers = NumDirectDrivers;	// The maximum number of sm
 
 #if SUPPORT_CAN_EXPANSION
 constexpr size_t MaxCanDrivers = 8;					// enough to support another Mini5+ as an expansion board plus a tool board, or eight closed-loop motors
-constexpr unsigned int MaxCanBoards = 8;			// enough to support eight closed loop motors
+constexpr size_t MaxCanBoards = 8;					// enough to support eight closed loop motors
 #endif
 
 constexpr size_t MaxPortsPerHeater = 2;
